--- recipe/openid/index.d.ts
@@ -1,10 +1,0 @@
-export * from "../../lib/build/recipe/openid";
-/**
- * 'export *' does not re-export a default.
- * import NextJS from "supertokens-node/nextjs";
- * the above import statement won't be possible unless either
- * - user add "esModuleInterop": true in their tsconfig.json file
- * - we do the following change:
- */
-import * as _default from "../../lib/build/recipe/openid";
-export default _default;
