--- recipe/openid/index.js
@@ -1,6 +1,0 @@
-"use strict";
-function __export(m) {
-    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
-}
-exports.__esModule = true;
-__export(require("../../lib/build/recipe/openid"));
