+++ test/test-server/src/index.ts
@@ -40,7 +40,6 @@ import ThirdParty from "../../../recipe/thirdparty";
 import TOTP from "../../../recipe/totp";
 import OAuth2 from "../../../recipe/oauth2";
 import OAuth2Client from "../../../recipe/oauth2client";
-import OpenId from "../../../recipe/openid";
 import accountlinkingRoutes from "./accountlinking";
 import emailpasswordRoutes from "./emailpassword";
 import emailverificationRoutes from "./emailverification";
@@ -93,7 +92,6 @@ function STReset() {
     TOTPRecipe.reset();
     OAuth2Recipe.reset();
     OAuth2ClientRecipe.reset();
-    OpenIdRecipe.reset();
     SuperTokensRecipe.reset();
 }
 
@@ -285,24 +283,6 @@ function initST(config: any) {
                 };
             }
             recipeList.push(OAuth2Client.init(initConfig));
-        }
-        if (recipe.recipeId === "openid") {
-            let initConfig: OpenIdRecipeTypeInput = {
-                ...config,
-            };
-            if (initConfig.override?.functions) {
-                initConfig.override = {
-                    ...initConfig.override,
-                    functions: getFunc(`${initConfig.override.functions}`),
-                };
-            }
-            if (initConfig.override?.apis) {
-                initConfig.override = {
-                    ...initConfig.override,
-                    apis: getFunc(`${initConfig.override.apis}`),
-                };
-            }
-            recipeList.push(OpenId.init(initConfig));
         }
     });
